Added OAuth along with the already present authentication

+++ bitbucket/client.py
@@ -1,4 +1,5 @@
 import requests
+import json
 
 from bitbucket.exceptions import UnknownError, InvalidIDError, NotFoundIDError, NotAuthenticatedError, PermissionError
 
@@ -6,7 +7,7 @@ from bitbucket.exceptions import UnknownError, InvalidIDError, NotFoundIDError, 
 class Client(object):
     BASE_URL = 'https://api.bitbucket.org/'
 
-    def __init__(self, user, password, owner=None):
+    def __init__(self, user=None, password=None, token=None):
         """Initial session with user/password, and setup repository owner
 
         Args:
@@ -18,13 +19,29 @@ class Client(object):
 
         self.user = user
         self.password = password
+        self.use_password = True
+        self.use_token = True
+        if not (user and password):
+            self.use_password = False
+        self.token = token
+        if not token:
+            self.use_token = False
 
+    def initialize(self, owner=None):
         user_data = self.get_user()
-
         # for shared repo, set baseURL to owner
-        if owner is None :
+        if owner is None:
             owner = user_data.get('username')
         self.username = owner
+
+    def get_token(self, client_id=None, client_secret=None):
+        if client_id and client_secret:
+            token_req_payload = {'grant_type': 'client_credentials'}
+            response = requests.post('https://bitbucket.org/site/oauth2/access_token', data=token_req_payload, allow_redirects=False, auth=(client_id, client_secret))
+            self.token = json.loads(response.text)['access_token']
+            self.use_token = True
+        else:
+            raise NotAuthenticatedError("Insufficient credentials")
 
     def get_user(self, params=None):
         """Returns the currently logged in user.
@@ -104,7 +121,6 @@ class Client(object):
         """
         return self._post('2.0/repositories/{}/{}'.format(team, name), params, data)
 
-    
     def get_repository_branches(self, repository_slug, params=None):
         return self._get('2.0/repositories/{}/{}/refs/branches'.format(self.username, repository_slug), params=params)
 
@@ -330,19 +346,75 @@ class Client(object):
                             params=params)
 
     def _get(self, endpoint, params=None):
-        response = requests.get(self.BASE_URL + endpoint, params=params, auth=(self.user, self.password))
+        if self.use_password:
+            response = requests.get(self.BASE_URL + endpoint, params=params, auth=(self.user, self.password))
+        elif self.use_token:
+            headers = {
+                "Accept": "application/json",
+                "Authorization": f"Bearer {self.token}"
+            }
+            response = requests.request(
+                "GET",
+                self.BASE_URL + endpoint,
+                params=params,
+                headers=headers
+            )
+        else:
+            raise NotAuthenticatedError("Insufficient credentials")
         return self._parse(response)
 
     def _post(self, endpoint, params=None, data=None):
-        response = requests.post(self.BASE_URL + endpoint, params=params, json=data, auth=(self.user, self.password))
+        if self.use_password:
+            response = requests.post(self.BASE_URL + endpoint, params=params, json=data, auth=(self.user, self.password))
+        elif self.use_token:
+            headers = {
+                "Accept": "application/json",
+                "Authorization": f"Bearer {self.token}"
+            }
+            response = requests.request(
+                "POST",
+                self.BASE_URL + endpoint,
+                params=params, json=data,
+                headers=headers
+            )
+        else:
+            raise NotAuthenticatedError("Insufficient credentials")
         return self._parse(response)
 
     def _put(self, endpoint, params=None, data=None):
-        response = requests.put(self.BASE_URL + endpoint, params=params, json=data, auth=(self.user, self.password))
+        if self.use_password:
+            response = requests.put(self.BASE_URL + endpoint, params=params, json=data, auth=(self.user, self.password))
+        elif self.use_token:
+            headers = {
+                "Accept": "application/json",
+                "Authorization": f"Bearer {self.token}"
+            }
+            response = requests.request(
+                "PUT",
+                self.BASE_URL + endpoint,
+                params=params, json=data,
+                headers=headers
+            )
+        else:
+            raise NotAuthenticatedError("Insufficient credentials")
         return self._parse(response)
 
     def _delete(self, endpoint, params=None):
-        response = requests.delete(self.BASE_URL + endpoint, params=params, auth=(self.user, self.password))
+        if self.use_password:
+            response = requests.delete(self.BASE_URL + endpoint, params=params, auth=(self.user, self.password))
+        elif self.use_token:
+            headers = {
+                "Accept": "application/json",
+                "Authorization": f"Bearer {self.token}"
+            }
+            response = requests.request(
+                "DELETE",
+                self.BASE_URL + endpoint,
+                params=params,
+                headers=headers
+            )
+        else:
+            raise NotAuthenticatedError("Insufficient credentials")
         return self._parse(response)
 
     def _parse(self, response):
